MAJOR UPDATE: Flipped algorithm around.
Running into scale problems after 100K rows. Hopefully this fixes it.

Essentially, before we were looping through files A and B, and storing the sum in a dict and searching for that sum when looping through file C.

This had two problems: it was extremely memory intensive - for even 10k row files, we’re storing 100M (10k^2) records at once. Additionally, because MPI requires objects to be pickled, I needed to keep the arrays as numpy arrays when sending the hashtable between processes. This took up a lot of memory and effeciency considering we didn’t actually need any operations on the arrays besides slicing.

Thus, the new approach is much better. To solve the second one, we’ve fully switched to memoryviews (instead of just one half).  Instead of storing the arrays themselves, we only need to keep the bytearray which is pickleable. Since we only needed to keep bytearrays, we don’t need any custom data structures (i.e. we can use a set). To solve the first one, we’ve switched the order of the algorithim.

The algorithim now works as follows:
1. Loop through file C, subtracting LAMBDA from each element in the array. The result, or difference, is the sum we need to find from the first two files.
2. Now loop through A and B, searching for the difference.