Add Ability To Exit After First Match
If we are only looking for one match, this change will allow the program to quit
immediately after finding one match. I opted to use an exit rather than a
nonblocking send to each worker because the way the program is set up, the MPI
interactions are separate from the three-way match code. Using non-blocking
sends, I would’ve had to include MPI into the three-way match code.

The improvement on complexity depends now on where the match is located. If the
first match is found at the end of the file, then the time improvement will be
negligible. However, if it’s found at the beginning or middle, there will be a
significant time improvement.